refactor: separate plotwidgets from views
